# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Hospital Management System.",
        "title": "HMS API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ekastn",
            "email": "eka@voixes.studio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5021",
    "basePath": "/api",
    "paths": {
        "/appointments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all appointments.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Get all appointments",
                "responses": {
                    "200": {
                        "description": "List of appointments",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.AppointmentDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve appointments",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new appointment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "parameters": [
                    {
                        "description": "Appointment object to be created",
                        "name": "appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Appointment created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create appointment",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/appointments/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a single appointment by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.AppointmentDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Appointment ID is required",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve appointment",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update details of an existing appointment.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Update an existing appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Appointment object with updated fields",
                        "name": "appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Appointment updated successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation failed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update appointment",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cancel an appointment (soft delete by changing status to 'Cancelled').",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Cancel an appointment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Appointment cancelled successfully",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Appointment ID is required",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to cancel appointment",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/appointments/{id}/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve detailed information for a single appointment, including patient and medical history.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Appointments"
                ],
                "summary": "Get detailed appointment information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Appointment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment details retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.AppointmentDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Appointment ID is required",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve appointment details",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.AppointmentDTO": {
            "type": "object",
            "required": [
                "dateTime",
                "doctorId",
                "duration",
                "location",
                "patientId",
                "status",
                "type"
            ],
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                },
                "dateTime": {
                    "type": "string",
                    "example": "2025-07-17T10:00:00Z"
                },
                "doctorId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000003"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "location": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Room 101"
                },
                "notes": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Patient complained of headache"
                },
                "patientHistory": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "No significant medical history"
                },
                "patientId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000002"
                },
                "status": {
                    "enum": [
                        "Scheduled",
                        "Confirmed",
                        "Completed",
                        "Cancelled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.AppointmentStatus"
                        }
                    ],
                    "example": "Scheduled"
                },
                "type": {
                    "enum": [
                        "check-up",
                        "follow-up",
                        "consultation",
                        "procedure",
                        "emergency"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.AppointmentType"
                        }
                    ],
                    "example": "check-up"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                }
            }
        },
        "domain.AppointmentDetailResponse": {
            "description": "Detailed appointment information",
            "type": "object",
            "properties": {
                "appointment": {
                    "$ref": "#/definitions/domain.AppointmentDTO"
                },
                "lastRecord": {
                    "$ref": "#/definitions/domain.MedicalRecordDTO"
                },
                "patient": {
                    "$ref": "#/definitions/domain.PatientDTO"
                }
            }
        },
        "domain.AppointmentStatus": {
            "type": "string",
            "enum": [
                "Scheduled",
                "Confirmed",
                "Completed",
                "Cancelled"
            ],
            "x-enum-varnames": [
                "AppointmentStatusScheduled",
                "AppointmentStatusConfirmed",
                "AppointmentStatusCompleted",
                "AppointmentStatusCancelled"
            ]
        },
        "domain.AppointmentType": {
            "type": "string",
            "enum": [
                "check-up",
                "follow-up",
                "consultation",
                "procedure",
                "emergency"
            ],
            "x-enum-varnames": [
                "AppointmentTypeCheckUp",
                "AppointmentTypeFollowUp",
                "AppointmentTypeConsultation",
                "AppointmentTypeProcedure",
                "AppointmentTypeEmergency"
            ]
        },
...
        "domain.MedicalRecordDTO": {
            "description": "Medical record data transfer object",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                },
                "date": {
                    "type": "string",
                    "example": "2025-07-17T10:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Patient presented with flu-like symptoms."
                },
                "diagnosis": {
                    "type": "string",
                    "example": "Influenza A"
                },
                "doctorId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000003"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "notes": {
                    "type": "string",
                    "example": "Advised patient to stay hydrated."
                },
                "patientId": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000002"
                },
                "recordType": {
                    "type": "string",
                    "example": "checkup"
                },
                "treatment": {
                    "type": "string",
                    "example": "Prescribed Tamiflu and rest."
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2025-07-17T09:00:00Z"
                }
            }
        },
        "domain.PatientDTO": {
            "description": "Patient data transfer object",
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "123 Main St"
                },
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string",
                    "example": "jane.doe@example.com"
                },
                "gender": {
                    "type": "string",
                    "example": "Female"
                },
                "id": {
                    "type": "string",
                    "example": "60d0fe4f53115a001f000001"
                },
                "lastVisit": {
                    "type": "string",
                    "example": "2025-07-17T10:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Jane Doe"
                },
                "phone": {
                    "type": "string",
                    "example": "1234567890"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}